+++ src/tool/subcommands/snapshot_cmd.rs
@@ -143,7 +143,7 @@ impl SnapshotCommands {
                     println!("{}", out.display());
                     Ok(())
                 }
-                Err(e) => cli_error_and_die(format!("Failed fetching the snapshot: {e}"), 1),
+                Err(e) => cli_error_and_die(format!("Failed fetching the snapshot: {e:#}"), 1),
             },
             Self::ValidateDiffs {
                 check_links,